first implementation of table card
