+++ src/app/app.component.html
@@ -1,3 +1,2 @@
-
 <app-navigation></app-navigation>
 <router-outlet></router-outlet>
