+++ src/app/material/material.module.ts
@@ -1,27 +1,30 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import {MatToolbarModule} from "@angular/material/toolbar";
-import {MatTabsModule} from "@angular/material/tabs";
-import {MatSidenavModule} from "@angular/material/sidenav";
-import {MatListModule} from "@angular/material/list";
-import {MatButtonModule} from "@angular/material/button";
-import {MatIconModule} from "@angular/material/icon";
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {MatToolbarModule} from '@angular/material/toolbar';
+import {MatTabsModule} from '@angular/material/tabs';
+import {MatSidenavModule} from '@angular/material/sidenav';
+import {MatListModule} from '@angular/material/list';
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule} from '@angular/material/icon';
+import {MatCardModule} from '@angular/material/card';
 
 const matModules = [
-  MatToolbarModule,
-  MatTabsModule,
-  MatSidenavModule,
-  MatListModule,
-  MatButtonModule,
-  MatIconModule
+    MatToolbarModule,
+    MatTabsModule,
+    MatSidenavModule,
+    MatListModule,
+    MatButtonModule,
+    MatCardModule,
+    MatIconModule
 ];
 
 @NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    ...matModules
-  ],
-  exports: [...matModules]
+    declarations: [],
+    imports: [
+        CommonModule,
+        ...matModules
+    ],
+    exports: [...matModules]
 })
-export class MaterialModule { }
+export class MaterialModule {
+}
