--- src/app/services/test-service.service.ts
@@ -1,20 +1,0 @@
-import { Injectable } from '@angular/core';
-import {HttpClient} from '@angular/common/http';
-import {Observable} from 'rxjs';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class TestServiceService {
-
-  constructor(private http: HttpClient) {
-  }
-
-  call1(): Observable<string>{
-    return this.http.get<string>('api/hello');
-  }
-
-  call2(): Observable<string>{
-    return this.http.get<string>('api/actuator/health');
-  }
-}
