+++ src/app/table-list/player-name.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { PlayerNamePipe } from './player-name.pipe';
+
+describe('PlayerNamePipe', () => {
+  it('create an instance', () => {
+    const pipe = new PlayerNamePipe();
+    expect(pipe).toBeTruthy();
+  });
+});
