+++ src/app/table-list/player-name.pipe.ts
@@ -1,0 +1,13 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import {Player} from './player.model';
+
+@Pipe({
+  name: 'playerName'
+})
+export class PlayerNamePipe implements PipeTransform {
+
+  transform(value: Player, ...args: any[]): any {
+    return value.firstName + ' ' + value.lastName;
+  }
+
+}
