+++ src/app/table-list/table-list.component.html
@@ -1,5 +1,4 @@
 <div class="tttable tttable--5cols">
-    table list works
     <div *ngFor="let currentTable of tables" class="tttable-cell">
         <app-tt-table [table]=currentTable></app-tt-table>
     </div>
