+++ src/app/table-list/tt-table/tt-table.component.html
@@ -1,1 +1,16 @@
-<p>{{table | json}}</p>
+<mat-card class="tt-table">
+    <mat-card-header>
+        <div mat-card-avatar class="example-header-image"></div>
+        <mat-card-title>Tisch Nr. {{table.id}}</mat-card-title>
+        <mat-card-subtitle>{{table.activeMatch.classification.name}}</mat-card-subtitle>
+    </mat-card-header>
+    <mat-card-content>
+        <div class="flex flex-horizontal-center card-main-center">
+            <div class="mat-body-strong">{{table.activeMatch.playerA | playerName}}</div>
+            <div>({{table.activeMatch.playerA.club.name}})</div>
+            <div>-</div>
+            <div class="mat-body-strong">{{table.activeMatch.playerB | playerName}}</div>
+            <div>({{table.activeMatch.playerB.club.name}})</div>
+        </div>
+    </mat-card-content>
+</mat-card>
