+++ src/app/table-list/tt-table/tt-table.component.scss
@@ -1,0 +1,9 @@
+.card-main-center {
+  text-align: center;
+  margin: 0 0 3em 0;
+  padding: 0;
+}
+
+.tt-table{
+  margin: 10px;
+}
