+++ src/app/table-list/tt-table/tt-table.component.ts
@@ -1,10 +1,10 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {Table} from "../table.model";
+import {Table} from '../table.model';
 
 @Component({
   selector: 'app-tt-table',
   templateUrl: './tt-table.component.html',
-  styleUrls: ['./tt-table.component.sass']
+  styleUrls: ['./tt-table.component.scss']
 })
 export class TtTableComponent implements OnInit {
